requirements.txt
fastapi
uvicorn
PyPDF2 
langchain==0.2.16
langchain-community==0.2.17 
python-dotenv
pinecone-client
cohere
groq==0.8.0
fastembed==0.3.6



